requirement.txt
streamlit
groq
python-dotenv
pandas
matplotlib
streamlit
groq
python-dotenv
pandas
matplotlib
PyPDF2